# Optimize performance of writing u128/i128 on 64bit system